+++ lib/rule-creation/game-rule-engine/src/InGameUserManager.cpp
@@ -1,8 +1,8 @@
 #include "InGameUserManager.h"
+#include "Server.h"
 #include <cassert>
 
-    // // Assume the role is passed as an Enum (stored within UserManager).
-    void InGameUserManager::addNewUser(uint32_t userID, Role role, GameEnvironment::Environment userStates){
+    void InGameUserManager::addNewUser(intptr_t userID, Role role, GameEnvironment::Environment userStates){
         auto iterator = m_userStates.find(userID);
 
         // // User should not already exist in the game.
@@ -13,7 +13,7 @@
         InGameUserManager::m_userStates.insert({userID, std::move(userStartingStates)});
     }
 
-    void InGameUserManager::deleteUser(uint32_t userID){
+    void InGameUserManager::deleteUser(intptr_t userID){
         // erase returns 0 if something was not erased.
         // If assert failed, that means the user already doesn't exist in this game
         // which should never happen if this is called.
@@ -21,20 +21,20 @@
     }
 
     // Ref: https://www.javatpoint.com/post/cpp-map-find-function
-    GameEnvironment::Environment InGameUserManager::getStatesOfUser(uint32_t userID){
+    GameEnvironment::Environment InGameUserManager::getStatesOfUser(intptr_t userID){
         auto iterator = m_userStates.find(userID);
         assert(iterator != m_userStates.end());
         std::pair<Role, GameEnvironment::Environment> statesToGet = std::move(iterator -> second);
         return std::move(statesToGet.second);
     }
 
-    std::map<uint32_t, std::pair<Role, GameEnvironment::Environment>> InGameUserManager::getAllUserStates(){
+    std::map<intptr_t, std::pair<Role, GameEnvironment::Environment>> InGameUserManager::getAllUserStates(){
         return std::move(InGameUserManager::m_userStates);
     }
 
     // // Recall that m_userStates is of type std::map<uint32_t, std::map<Role, Environment>>
     // // Create a new pair to replace the existing one that the User ID maps to.
-    void InGameUserManager::setStatesOfUser(uint32_t userID, Role role, GameEnvironment::Environment states){
+    void InGameUserManager::setStatesOfUser(intptr_t userID, Role role, GameEnvironment::Environment states){
         auto iterator = m_userStates.find(userID);
         // User SHOULD already exist in the game.
         assert(iterator != m_userStates.end());
